Renamed CallNode.swift file to CallMacroNodes.swift + splitting Filters in 3 files
